// src/components/dashBoard/types/returnData.ts
export interface statisticsData {
    totalUsers: number
    newUsersToday: number
    newUsersThisMonth: number
    activeUsersToday: number
    totalReservations: number
    reservationsToday: number
    reservationsThisMonth: number
    reservationRevenue: number
    revenueToday: number
    revenueThisMonth: number
    totalOrders: number
    ordersToday: number
    ordersThisMonth: number
    mallRevenue: number
    mallRevenueToday: number
    mallRevenueThisMonth: number
    totalPosts: number
    postsToday: number
    totalReplies: number
    repliesToday: number
    totalVenues: number
    availableVenues: number
    venueUtilizationRate: number
}